01. LangChain-RAG/requirements.txt
langchain
langchain_core
langchain_community
langchain_openai
streamlit
dotenv
pdfminer.six
faiss-cpu